refactor: retire unconsumed slotted runtime wrapper
