+++ crates/noon-runtime/src/execution_slots/mod.rs
@@ -1,9 +1,9 @@
-mod base;
 mod runtime_transaction;
 mod semantic_membership;
+mod table;
 
-pub use base::*;
 pub use runtime_transaction::{
     AuthoredPublicationError, PreparedAuthoredPlanChange, PreparedAuthoredReactivePlanChange,
 };
 pub use semantic_membership::*;
+pub use table::*;
